警告レベル4への引き上げと警告削除
デバッグ機能の強化とエラーチェックの追加

+++ Engine/Features/UI/UINavigator.cpp
@@ -130,7 +130,7 @@ Direction UINavigator::GetDirectionFromInput() const
     {
         for (const auto& keycode : keys)
         {
-            if (input_->IsPadTriggered(static_cast<PadButton>(keycode)) || input_->IsKeyTriggered(keycode))
+            if (input_->IsPadTriggered(static_cast<PadButton>(static_cast<uint8_t>(keycode))) || input_->IsKeyTriggered(static_cast<uint8_t>(keycode)))
             {
                 switch (action)
                 {
